backend/requirements.txt
fastapi
uvicorn[standard]
google-generativeai
resend
httpx
pydantic
pydantic-settings
python-dotenv
